Don't finish statements with ";"
This exposes a problem or difference/lack of functionality in
DBI::SQL::Nano which sees a trailing ';' as part of a bad ORDER BY
statement. I think I only added it for readability... Possibly this
should be raised as an issue in SQL::Nano? But in the meantime I don't
want my tests to fail for that.