Minor syntax improvement for UNION statements
My preferred way of writing union statements is this:

    select    => 1,
    union_all => '',
    select    => 2,

Which with this change now looks like this:

    SELECT
        1

    UNION ALL

    SELECT
        2
    ;

This change merely adds the extra newline between the union and the
previous select statement to give selects a bit of space. Looks a bit
trivial for this basic example, but with large queries it makes a
decent visual difference.